Removed AbstractComponentDataProvider
