+++ Addons/Entitas.CodeGeneration.Plugins/Entitas.CodeGeneration.Plugins/Plugins/DataProviders/Components/ComponentDataProvider.cs
@@ -1,20 +1,17 @@
-using System;
+﻿using System;
+using System.Linq;
+using Entitas.CodeGeneration.Attributes;
+using Entitas.CodeGeneration.CodeGenerator;
+using Entitas.Utils;
 
 namespace Entitas.CodeGeneration.Plugins {
 
-    public class ComponentDataProvider : AbstractComponentDataProvider {
+    public class ComponentDataProvider : ICodeGeneratorDataProvider {
 
-        public override string name { get { return "Component"; } }
-        public override int priority { get { return 0; } }
-        public override bool isEnabledByDefault { get { return true; } }
-        public override bool runInDryMode { get { return true; } }
-
-        public ComponentDataProvider() : this(null) {
-        }
-
-        public ComponentDataProvider(Type[] types)
-            : base(types, getComponentDataProviders()) {
-        }
+        public string name { get { return "Component"; } }
+        public int priority { get { return 0; } }
+        public bool isEnabledByDefault { get { return true; } }
+        public bool runInDryMode { get { return true; } }
 
         static IComponentDataProvider[] getComponentDataProviders() {
             return new IComponentDataProvider[] {
@@ -27,6 +24,82 @@ namespace Entitas.CodeGeneration.Plugins {
                 new ShouldGenerateMethodsComponentDataProvider(),
                 new ShouldGenerateComponentIndexComponentDataProvider()
             };
+        }
+
+        Type[] _types;
+        IComponentDataProvider[] _dataProviders;
+
+        public ComponentDataProvider() : this(null) {
+        }
+
+        public ComponentDataProvider(Type[] types) : this(types, getComponentDataProviders()) {
+        }
+
+        protected ComponentDataProvider(Type[] types, IComponentDataProvider[] dataProviders) {
+            _types = types;
+            _dataProviders = dataProviders;
+        }
+
+        public CodeGeneratorData[] GetData() {
+            if(_types == null) {
+                _types = CodeGeneratorUtil.LoadTypesFromAssemblies();
+            }
+
+            var dataFromComponents = _types
+                .Where(type => type.ImplementsInterface<IComponent>())
+                .Where(type => !type.IsAbstract)
+                .Select(type => createDataForComponent(type));
+
+            var dataFromNonComponents = _types
+                .Where(type => !type.ImplementsInterface<IComponent>())
+                .Where(type => !type.IsGenericType)
+                .Where(type => hasContexts(type))
+                .SelectMany(type => createDataForNonComponent(type));
+
+            var generatedComponentsLookup = dataFromNonComponents.ToLookup(data => data.GetFullTypeName());
+            return dataFromComponents
+                .Where(data => !generatedComponentsLookup.Contains(data.GetFullTypeName()))
+                .Concat(dataFromNonComponents)
+                .ToArray();
+        }
+
+        ComponentData createDataForComponent(Type type) {
+            var data = new ComponentData();
+            foreach(var provider in _dataProviders) {
+                provider.Provide(type, data);
+            }
+
+            return data;
+        }
+
+        ComponentData[] createDataForNonComponent(Type type) {
+            return getComponentNames(type)
+                .Select(componentName => {
+                    var data = createDataForComponent(type);
+                    data.SetFullTypeName(componentName.AddComponentSuffix());
+                    data.SetMemberData(new[] {
+                        new MemberData(type.ToCompilableString(), "value")
+                    });
+
+                    return data;
+                }).ToArray();
+        }
+
+        bool hasContexts(Type type) {
+            return ContextsComponentDataProvider.GetContextNames(type).Length != 0;
+        }
+
+        string[] getComponentNames(Type type) {
+            var attr = Attribute
+                .GetCustomAttributes(type)
+                .OfType<CustomComponentNameAttribute>()
+                .SingleOrDefault();
+
+            if(attr == null) {
+                return new[] { type.ToCompilableString().ShortTypeName().AddComponentSuffix() };
+            }
+
+            return attr.componentNames;
         }
     }
 }
